chapter01 JUnitチュートリアル：除算メソッド少数に対応（テスト駆動開発）

+++ JUnit-Tutorial/src/main/java/com/slgerkamp/junit/chapter01/Calculator.java
@@ -4,7 +4,7 @@ public class Calculator {
 	public int multiply(int x, int y){
 		return x * y;
 	}
-	public int divide(int x, int y){
-		return x / y;
+	public float divide(int x, int y){
+		return (float) x / (float) y;
 	}
 }
